@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '>=3.2.4 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,12 +30,19 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-#  rive: ^0.13.9
+  #  rive: ^0.13.9
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  freezed_annotation: ^2.4.4
+  json_annotation: ^4.9.0
+  rive: ^0.13.4
+  dio: ^5.5.0+1
+  fl_chart: ^0.68.0
+  url_launcher: ^6.3.0
+  http: ^1.2.0
 
 dev_dependencies:
   flutter_test:
@@ -47,6 +54,9 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^2.0.0
+  build_runner: ^2.4.9
+  freezed: ^2.5.2
+  json_serializable: ^6.8.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -60,7 +70,7 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
+  #  assets:
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
 
