@@ -1,6 +1,13 @@
 class Ingredient {
+  final id;
   final name;
   final count;
+  final unit;
 
-  Ingredient({required this.name, required this.count});
+  Ingredient({
+    required this.id,
+    required this.name,
+    required this.count,
+    required this.unit,
+  });
 }
