@@ -1,21 +1,21 @@
 class Product {
   final int id;
   final String name;
-  final int serving;
-  final int? protein;
-  final int? fat;
-  final int? carbohydrates;
-  final int? energyValue;
-  final String? categories;
+  final double serving;
+  final double proteins;
+  final double fats;
+  final double carbohydrates;
+  final double calories;
+  final String categories;
 
   Product({
     required this.id,
     required this.name,
     required this.serving,
-    this.protein,
-    this.fat,
-    this.carbohydrates,
-    this.energyValue,
-    this.categories,
+    required this.proteins,
+    required this.fats,
+    required this.carbohydrates,
+    required this.calories,
+    required this.categories,
   });
 }
