@@ -1,12 +1,13 @@
 class TotalParams {
-  final int totalCalories;
-  final int totalFats;
-  final int totalProteins;
-  final int totalCarbohydrate;
-  final int requiredCaloties;
-  final int dailyProteinNeeds;
-  final int dailyFatNeeds;
-  final int dailyCarbohydrateNeeds;
+  final double totalCalories;
+  final double totalFats;
+  final double totalProteins;
+  final double totalCarbohydrate;
+  final double requiredCaloties;
+  final double dailyProteinNeeds;
+  final double dailyFatNeeds;
+  final double dailyCarbohydrateNeeds;
+  final double imt;
 
   TotalParams({
     required this.totalCalories,
@@ -17,5 +18,6 @@ class TotalParams {
     required this.dailyProteinNeeds,
     required this.dailyFatNeeds,
     required this.dailyCarbohydrateNeeds,
+    required this.imt,
   });
 }
